feat: library archives - LegendaryRevolt feat: made damage resistance int > float

+++ src/main/java/com/ebicep/warlords/abilities/CapacitorTotem.java
@@ -148,7 +148,7 @@ public class CapacitorTotem extends AbstractTotem implements Duration {
                             });
 
                             if (pveMasterUpgrade) {
-                                int damageResistance = warlordsPlayer.getSpec().getDamageResistance();
+                                float damageResistance = warlordsPlayer.getSpec().getDamageResistance();
                                 warlordsPlayer.setDamageResistance(damageResistance - 20);
                             }
                         });
